mawrid_user/requirements.txt
pytest-html>=3.2.0